Breakout middleware into separate files
