+++ src/stac_auth_proxy/middleware/__init__.py
@@ -1,0 +1,11 @@
+"""Custom middleware."""
+
+from .OpenApiMiddleware import OpenApiMiddleware
+from .AddProcessTimeHeaderMiddleware import AddProcessTimeHeaderMiddleware
+from .EnforceAuthMiddleware import EnforceAuthMiddleware
+
+__all__ = [
+    OpenApiMiddleware,
+    AddProcessTimeHeaderMiddleware,
+    EnforceAuthMiddleware,
+]
